requirements.txt
appdirs==1.4.4
asgiref==3.3.4
black==21.5b1
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
Django==3.2.3
django-cors-headers==3.7.0
django-debug-toolbar==3.2.1
django-decouple==2.1
djangorestframework==3.12.4
gunicorn==20.1.0
idna==2.10
importlib-metadata==4.5.0
mypy-extensions==0.4.3
mysqlclient==2.0.3
pathspec==0.8.1
pytz==2021.1
regex==2021.4.4
requests==2.25.1
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.3
typing==3.7.4.3
typing-extensions==3.10.0.0
urllib3==1.26.5
zipp==3.4.1

pip~=18.1
colorama~=0.4.4
PyYAML~=6.0
setuptools~=40.6.2
tablib~=3.1.0
python-dotenv~=0.19.2
xlrd==2.0.1
XlsxWriter==3.0.2
xlwt==1.3.0
zipp==3.4.1
pika==1.1.0

